refactor: ATLAS-NONE: Refactor HlaMetadata service.
