+++ Atlas.HlaMetadataDictionary.Test/TestHelpers/Builders/HlaMetadataDictionaryBuilder.cs
@@ -17,7 +17,7 @@ namespace Atlas.HlaMetadataDictionary.Test.TestHelpers.Builders
         private ILocusHlaMatchingMetadataService locus;
         private IHlaScoringMetadataService scoring;
         private IDpb1TceGroupMetadataService dpb1;
-        private IHlaMetadataService metadata;
+        private IHlaMetadataGenerationOrchestrator metadata;
         private IWmdaHlaNomenclatureVersionAccessor wmdaHlaNomenclatureVersionAccessor;
         private ILogger logger;
         private IHlaMetadataDictionary cannedResponse = null;
@@ -35,7 +35,7 @@ namespace Atlas.HlaMetadataDictionary.Test.TestHelpers.Builders
             locus = Substitute.For<ILocusHlaMatchingMetadataService>();
             scoring = Substitute.For<IHlaScoringMetadataService>();
             dpb1 = Substitute.For<IDpb1TceGroupMetadataService>();
-            metadata = Substitute.For<IHlaMetadataService>();
+            metadata = Substitute.For<IHlaMetadataGenerationOrchestrator>();
             wmdaHlaNomenclatureVersionAccessor = Substitute.For<IWmdaHlaNomenclatureVersionAccessor>();
             logger = Substitute.For<ILogger>();
         }
@@ -68,7 +68,7 @@ namespace Atlas.HlaMetadataDictionary.Test.TestHelpers.Builders
                 case IDpb1TceGroupMetadataService typedDependency:
                     dpb1 = typedDependency;
                     break;
-                case IHlaMetadataService typedDependency:
+                case IHlaMetadataGenerationOrchestrator typedDependency:
                     metadata = typedDependency;
                     break;
                 case IWmdaHlaNomenclatureVersionAccessor typedDependency:
