+++ Atlas.HlaMetadataDictionary.Test/UnitTests/Services/DataGeneration/MatchedHlaConversion/HlaToScoringMetaDataConverterTests.cs
@@ -8,7 +8,7 @@ using Atlas.HlaMetadataDictionary.Services.DataGeneration.MatchedHlaConversion;
 using FluentAssertions;
 using NUnit.Framework;
 
-namespace Atlas.HlaMetadataDictionary.Test.UnitTests.Services.DataRetrieval.MatchedHlaConversion
+namespace Atlas.HlaMetadataDictionary.Test.UnitTests.Services.DataGeneration.MatchedHlaConversion
 {
     [TestFixture]
     internal class HlaToScoringMetaDataConverterTests :
