+++ Atlas.HlaMetadataDictionary/ExternalInterface/DependencyInjection/ServiceConfiguration.cs
@@ -77,7 +77,7 @@ namespace Atlas.HlaMetadataDictionary.ExternalInterface.DependencyInjection
             services.AddScoped<IHlaConverter, HlaConverter>();
             services.AddScoped<IHlaNameToTwoFieldAlleleConverter, HlaNameToTwoFieldAlleleConverter>();
             services.AddScoped<IAlleleNamesMetadataService, AlleleNamesMetadataService>();
-            services.AddScoped<IHlaMetadataService, HlaMetadataService>();
+            services.AddScoped<IHlaMetadataGenerationOrchestrator, HlaMetadataGenerationOrchestrator>();
             services.AddScoped<ILocusHlaMatchingMetadataService, LocusHlaMatchingMetadataService>();
             services.AddScoped<IHlaMatchingMetadataService, HlaMatchingMetadataService>();
             services.AddScoped<IHlaScoringMetadataService, HlaScoringMetadataService>();
