+++ Atlas.HlaMetadataDictionary/ExternalInterface/HlaMetadataDictionary.cs
@@ -50,7 +50,7 @@ namespace Atlas.HlaMetadataDictionary.ExternalInterface
         private readonly ILocusHlaMatchingMetadataService locusHlaMatchingMetadataService;
         private readonly IHlaScoringMetadataService hlaScoringMetadataService;
         private readonly IDpb1TceGroupMetadataService dpb1TceGroupMetadataService;
-        private readonly IHlaMetadataService hlaMetadataService;
+        private readonly IHlaMetadataGenerationOrchestrator hlaMetadataGenerationOrchestrator;
         private readonly IWmdaHlaNomenclatureVersionAccessor wmdaHlaNomenclatureVersionAccessor;
         private readonly ILogger logger;
 
@@ -62,7 +62,7 @@ namespace Atlas.HlaMetadataDictionary.ExternalInterface
             ILocusHlaMatchingMetadataService locusHlaMatchingMetadataService,
             IHlaScoringMetadataService hlaScoringMetadataService,
             IDpb1TceGroupMetadataService dpb1TceGroupMetadataService,
-            IHlaMetadataService hlaMetadataService,
+            IHlaMetadataGenerationOrchestrator hlaMetadataGenerationOrchestrator,
             IWmdaHlaNomenclatureVersionAccessor wmdaHlaNomenclatureVersionAccessor,
             ILogger logger)
         {
@@ -73,7 +73,7 @@ namespace Atlas.HlaMetadataDictionary.ExternalInterface
             this.locusHlaMatchingMetadataService = locusHlaMatchingMetadataService;
             this.hlaScoringMetadataService = hlaScoringMetadataService;
             this.dpb1TceGroupMetadataService = dpb1TceGroupMetadataService;
-            this.hlaMetadataService = hlaMetadataService;
+            this.hlaMetadataGenerationOrchestrator = hlaMetadataGenerationOrchestrator;
             this.wmdaHlaNomenclatureVersionAccessor = wmdaHlaNomenclatureVersionAccessor;
             this.logger = logger;
         }
@@ -162,7 +162,7 @@ namespace Atlas.HlaMetadataDictionary.ExternalInterface
 
         public HlaMetadataCollection GetAllHlaMetadata(string version)
         {
-            return hlaMetadataService.GetAllHlaMetadata(version);
+            return hlaMetadataGenerationOrchestrator.GenerateAllHlaMetadata(version);
         }
     }
 }
